2017-08-01 (basic model) 	- basicModel.households: paramétrisation des comportements 	des ménages. 	- data.DataManager: données calculées sur des intervalles de 	plusieurs périodes (sommes). 	- data.DataManager: mise en place d'un pool pour les 	expressions ainsi que de caches pour leurs résultats. 	- gui.BasicGui: introduction d'un Timer qui consulte à 	intervales réguliers l'avancement de la simulation et gère 	la mise à jour de l'interface. 	- gui.JamelChartFactory: axes affichés par défaut. 	- gui.JamelChartFactory: Ajout d'une légende des graphes, 	- util.BasicSector: l'ouverture et la fermeture des agents 	sont gérées automatiquemet par le secteur (au lieu que ce 	soit une phase). 	- util.Parameters: nouvel objet qui encapsule les paramètres 	et facilite le parsing des fichiers XML.

+++ src/jamel/basicModel/firms/BasicFactory.java
@@ -3,7 +3,7 @@ package jamel.basicModel.firms;
 import java.util.List;
 
 import jamel.Jamel;
-import jamel.util.AgentDataset;
+import jamel.data.AgentDataset;
 import jamel.util.BasicTimer;
 import jamel.util.JamelObject;
 import jamel.util.Timer;
@@ -82,7 +82,7 @@ class BasicFactory extends JamelObject implements Factory {
 	}
 
 	@Override
-	public double getCapacity() {
+	public int getCapacity() {
 		return this.capacity;
 	}
